make scenarios and PR enchancement
