+++ .github/workflows/master.yaml
@@ -1,15 +1,7 @@
-# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
-name: ls-branch:build
+name: label-studio:build
 
 on:
   push:
-    branches:
-      - '*'
-      - '!master'
-
-  pull_request:
     branches: [ master ]
 
 jobs:
@@ -39,3 +31,4 @@ jobs:
         python -m pytest -vrP
         coverage run -m --source=label_studio pytest
         coverage report -m
+
